requirements.txt
Pillow>=11.1.0
svglib>=1.5.1
reportlab>=4.3.1
